requirements.txt
# Clinical AI Data Loading Module - Python Dependencies
# Tested with Python 3.10.12. Other Python versions are currently untested.

# Core ML/Data Science Stack
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# Image Processing
Pillow==10.0.0
opencv-python==4.8.0.76

# Configuration and Utilities
PyYAML==6.0
python-dotenv==1.0.0

# Development Tools (optional but recommended)
pytest==7.4.0
black==23.7.0
flake8==6.0.0
mypy==1.4.1

# Experiment Tracking (optional)
# wandb==0.15.9
# tensorboard==2.13.0

# Additional utilities
tqdm==4.65.0
matplotlib==3.7.2
